Clean up applicative regexes
