+++ regex/src/main/scala/applicative/Greediness.scala
@@ -1,0 +1,10 @@
+package ceedubs.irrec
+package regex
+package applicative
+
+sealed abstract class Greediness extends Product with Serializable
+
+object Greediness {
+  case object Greedy extends Greediness
+  case object NonGreedy extends Greediness
+}
