+++ tests/src/test/scala/regex/applicative/RETests.scala
@@ -1,0 +1,10 @@
+package ceedubs.irrec
+package regex
+package applicative
+
+import ceedubs.irrec.regex.Regex._
+
+// TODO
+class RETests extends IrrecSuite {
+  test("literal match") { assert(literal('b').stringMatcher("b")) }
+}
